+++ crates/medea-reactive/src/collections/hash_map.rs
@@ -8,12 +8,13 @@ use std::{
     marker::PhantomData,
 };
 
-use futures::{
-    future::{self, FutureExt as _, LocalBoxFuture},
-    stream::LocalBoxStream,
-};
+use futures::stream::{LocalBoxStream, StreamExt as _};
 
-use crate::subscribers_store::{common, progressable, SubscribersStore};
+use crate::subscribers_store::{
+    common, progressable,
+    progressable::{AllProcessed, Processed},
+    SubscribersStore,
+};
 
 /// Reactive hash map based on [`HashMap`][1] with additional functionality of
 /// tracking progress made by its subscribers. Its [`HashMap::on_insert()`] and
@@ -122,8 +123,7 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_insert_processed(&self) -> LocalBoxFuture<'static, ()> {
+    pub fn when_insert_processed(&self) -> Processed<'static, ()> {
         self.on_insert_subs.when_all_processed()
     }
 
@@ -132,8 +132,7 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_remove_processed(&self) -> LocalBoxFuture<'static, ()> {
+    pub fn when_remove_processed(&self) -> Processed<'static, ()> {
         self.on_remove_subs.when_all_processed()
     }
 
@@ -142,15 +141,11 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_all_processed(&self) -> LocalBoxFuture<'static, ()> {
-        Box::pin(
-            future::join(
-                self.when_remove_processed(),
-                self.when_insert_processed(),
-            )
-            .map(|(_, _)| ()),
-        )
+    pub fn when_all_processed(&self) -> AllProcessed<'static, ()> {
+        crate::when_all_processed(vec![
+            self.when_remove_processed().into(),
+            self.when_insert_processed().into(),
+        ])
     }
 }
 
@@ -222,6 +217,12 @@ where
                 .map(|(k, v)| self.on_insert_subs.wrap((k.clone(), v.clone())))
                 .collect::<Vec<_>>(),
         ))
+    }
+
+    /// Chains [`HashMap::replay_on_insert()`] with a [`HashMap::on_insert()`].
+    #[inline]
+    pub fn on_insert_with_replay(&self) -> LocalBoxStream<'static, O> {
+        Box::pin(self.replay_on_insert().chain(self.on_insert()))
     }
 }