+++ crates/medea-reactive/src/collections/vec.rs
@@ -2,11 +2,13 @@
 
 use std::{marker::PhantomData, slice::Iter};
 
-use futures::{
-    future, future::LocalBoxFuture, stream::LocalBoxStream, FutureExt as _,
-};
+use futures::stream::LocalBoxStream;
 
-use crate::subscribers_store::{common, progressable, SubscribersStore};
+use crate::subscribers_store::{
+    common, progressable,
+    progressable::{processed::AllProcessed, Processed},
+    SubscribersStore,
+};
 
 /// Reactive vector based on [`Vec`] with additional functionality of tracking
 /// progress made by its subscribers. Its [`Vec::on_push()`] and
@@ -102,8 +104,7 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_push_processed(&self) -> LocalBoxFuture<'static, ()> {
+    pub fn when_push_processed(&self) -> Processed<'static, ()> {
         self.on_push_subs.when_all_processed()
     }
 
@@ -112,8 +113,7 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_remove_processed(&self) -> LocalBoxFuture<'static, ()> {
+    pub fn when_remove_processed(&self) -> Processed<'static, ()> {
         self.on_remove_subs.when_all_processed()
     }
 
@@ -122,15 +122,11 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_all_processed(&self) -> LocalBoxFuture<'static, ()> {
-        Box::pin(
-            future::join(
-                self.when_remove_processed(),
-                self.when_push_processed(),
-            )
-            .map(|(_, _)| ()),
-        )
+    pub fn when_all_processed(&self) -> AllProcessed<'static, ()> {
+        crate::when_all_processed(vec![
+            self.when_remove_processed().into(),
+            self.when_push_processed().into(),
+        ])
     }
 }